refactor: FavoriteController 구축
