+++ src/main/java/nextstep/subway/member/ui/FavoriteController.java
@@ -1,0 +1,60 @@
+package nextstep.subway.member.ui;
+
+import nextstep.subway.auth.application.AuthorizationException;
+import nextstep.subway.auth.domain.AuthenticationPrincipal;
+import nextstep.subway.auth.domain.LoginMember;
+import nextstep.subway.member.application.MemberService;
+import nextstep.subway.member.domain.Favorite;
+import nextstep.subway.member.dto.FavoriteRequest;
+import nextstep.subway.member.dto.FavoriteResponse;
+import nextstep.subway.station.application.StationService;
+import nextstep.subway.station.domain.Stations;
+
+import java.net.URI;
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/favorites")
+public class FavoriteController {
+    private MemberService memberService;
+    private StationService stationService;
+
+    public FavoriteController(MemberService memberService, StationService stationService) {
+        this.memberService = memberService;
+        this.stationService = stationService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> createFavorite(@AuthenticationPrincipal LoginMember loginMember, @RequestBody FavoriteRequest request) {
+        Stations stations = new Stations(stationService.findByIds(request.toStationIds()));
+        Favorite favorite = memberService.createFavorite(loginMember.getId(), stations.getStation(request.getSource()), stations.getStation(request.getTarget()));
+        return ResponseEntity.created(URI.create("/favorites/" + favorite.getId())).build();
+    }
+
+    @GetMapping
+    public ResponseEntity<List<FavoriteResponse>> findFavorites(@AuthenticationPrincipal LoginMember loginMember) {
+        return ResponseEntity.ok().body(memberService.findFavorites(loginMember.getId()));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> removeFavorite(@AuthenticationPrincipal LoginMember loginMember, @PathVariable("id") Long favoriteId) {
+        memberService.removeFavorite(loginMember.getId(), favoriteId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @ExceptionHandler(AuthorizationException.class)
+    public ResponseEntity<String> handleAuthorizationException(Exception e) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(e.getMessage());
+    }
+}
