+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -6,16 +6,10 @@ import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.auth.infrastructure.JwtTokenProvider;
 import nextstep.subway.member.application.MemberService;
-import nextstep.subway.member.domain.Favorite;
-import nextstep.subway.member.dto.FavoriteRequest;
-import nextstep.subway.member.dto.FavoriteResponse;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
-import nextstep.subway.station.application.StationService;
-import nextstep.subway.station.domain.Stations;
 
 import java.net.URI;
-import java.util.List;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -32,12 +26,10 @@ import org.springframework.web.bind.annotation.RestController;
 public class MemberController {
     private MemberService memberService;
     private JwtTokenProvider jwtTokenProvider;
-    private StationService stationService;
 
-    public MemberController(MemberService memberService, JwtTokenProvider jwtTokenProvider, StationService stationService) {
+    public MemberController(MemberService memberService, JwtTokenProvider jwtTokenProvider) {
         this.memberService = memberService;
         this.jwtTokenProvider = jwtTokenProvider;
-        this.stationService = stationService;
     }
 
     @PostMapping("/members")
@@ -80,24 +72,6 @@ public class MemberController {
     @DeleteMapping("/members/me")
     public ResponseEntity<MemberResponse> deleteMemberOfMine(@AuthenticationPrincipal LoginMember loginMember) {
         memberService.deleteMember(loginMember.getId());
-        return ResponseEntity.noContent().build();
-    }
-
-    @PostMapping("/favorites")
-    public ResponseEntity<Void> createFavorite(@AuthenticationPrincipal LoginMember loginMember, @RequestBody FavoriteRequest request) {
-        Stations stations = new Stations(stationService.findByIds(request.toStationIds()));
-        Favorite favorite = memberService.createFavorite(loginMember.getId(), stations.getStation(request.getSource()), stations.getStation(request.getTarget()));
-        return ResponseEntity.created(URI.create("/favorites/" + favorite.getId())).build();
-    }
-
-    @GetMapping("/favorites")
-    public ResponseEntity<List<FavoriteResponse>> findFavorites(@AuthenticationPrincipal LoginMember loginMember) {
-        return ResponseEntity.ok().body(memberService.findFavorites(loginMember.getId()));
-    }
-
-    @DeleteMapping("/favorites/{id}")
-    public ResponseEntity<Void> removeFavorite(@AuthenticationPrincipal LoginMember loginMember, @PathVariable("id") Long favoriteId) {
-        memberService.removeFavorite(loginMember.getId(), favoriteId);
         return ResponseEntity.noContent().build();
     }
